# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all transactions",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Record transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get transaction by ID",
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update transaction",
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/reviews": {
      "get": {
        "operationId": "ReviewsController_findAll",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all reviews with stats",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}/reply": {
      "post": {
        "operationId": "ReviewsController_reply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Reply to a student review",
        "tags": [
          "Reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_deleteReply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete reply to a student review",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coursera App API",
    "description": "The API for Admin and Instructor modules",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
...
  }
}